Feature(Loss): Added two loss functions. (#3)
* feat: Added two type of loss functions.

* feat: Changed name.

* feat: Fixed loss functions, cleaned up code, added type annotations.

* feat: Added embedding and test cases.

* feat: Added more verbose test print.

+++ parse.py
@@ -28,7 +28,7 @@ def parse_program(input: str) -> list[Instruction]:
 
 
 def parse_points(input: str) -> list[Point]:
-    return [tuple(map(int, line.split(' '))) for line in input.splitlines()]
+    return [tuple(map(float, line.split(' '))) for line in input.splitlines()]
 
 program = parse_program('CIRCLE 2.1\nTRANSLATION 1.2 3.4')
 print(program)
